refactor(llcppsymg):with default dylib fetch
